requirements.txt
requests==2.28.1
pytest==7.2.0
vcrpy==4.2.1